Refactor involvement api fetches
