--- modules/shows-fetch.js
@@ -1,5 +1,0 @@
-export default function fetchShow(endpoint, id, embed = '') {
-  const fetching = fetch(`https://api.tvmaze.com/${endpoint}/${id}${embed}`)
-    .then((response) => response.json());
-  return fetching;
-}
